// src/gaos/models/security.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

export type Security = {
  /**
   * OAuth access token sent as a bearer Authorization header.
   */
  access_token?: string | undefined;
  /**
   * Gemini API key sent as x-goog-api-key.
   */
  api_key?: string | undefined;
  /**
   * Additional default headers to apply before request-specific headers and auth.
   */
  default_headers?: { [k: string]: string } | undefined;
};